+++ activerecord/CHANGELOG.md
@@ -1,1 +1,8 @@
+*  `Migrator::migrations_paths` will now use the application's `db/migrate` paths
+    instead of assuming `'db/migrate'`.
+
+    Fixes #22261.
+
+    *Nate Salisbury*
+
 Please check [5-1-stable](https://github.com/rails/rails/blob/5-1-stable/activerecord/CHANGELOG.md) for previous changes.